+++ src/fmt/writer/wasm.rs
